M5L4/VISION-AVES/requirements.txt
flask
tensorflow>=2.12
pillow
numpy